requirements.txt
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
narwhals==1.27.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
packaging==24.2
pillow==11.1.0
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
